requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-markdownx==4.0.7
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
